requirements.txt
discord.py>=2.3.0
colorama>=0.4.6
psutil>=5.9.5
python-dotenv>=1.0.0
aiohttp>=3.8.5
 
